requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
numpy
huggingface_hub
python-dotenv 